% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpand.R
\name{filterDates}
\alias{filterDates}
\title{Filter archaeological site coordinates and dates, retaining only the
earliest radiocarbon date per site or per region if a radius is specified.}
\usage{
filterDates(sites, c14bp, dist = 0)
}
\arguments{
\item{sites}{A SpatialPointsDataFrame object with archaeological sites and
associated radiocarbon ages.}

\item{c14bp}{A string. Name of the field with the radiocarbon ages in C14 BP
format.}

\item{dist}{A number. Radius in km to filter for the earliest dates in a
region. Default is 0, i.e. the earliest dates of all sites are retained.}
}
\value{
A SpatialPointsDataFrame object with the earliest C14 date for every
site.
}
\description{
Filter archaeological site coordinates and dates, retaining only the
earliest radiocarbon date per site or per region if a radius is specified.
}
