% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpand.R
\name{interpolateIDW}
\alias{interpolateIDW}
\title{Interpolate radiocarbon dates using inverse distance weighting.}
\usage{
interpolateIDW(sites, c14bp)
}
\arguments{
\item{sites}{A SpatialPointsDataFrame object with archaeological sites and
associated radiocarbon ages.}

\item{c14bp}{A string. Name of the field with the radiocarbon ages in C14 BP
format.}
}
\value{
A RasterLayer with interpolated C14 ages.
}
\description{
Interpolate radiocarbon dates using inverse distance weighting.
}
