% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpand.R
\name{isoPlot}
\alias{isoPlot}
\title{Plots an interpolated surface with the radiocarbon ages of archaeological
sites.}
\usage{
isoPlot(sites, c14bp, title = "All sites")
}
\arguments{
\item{sites}{A SpatialPointsDataFrame object.}

\item{c14bp}{A string. Name of the field with the radiocarbon ages in C14 BP
format.}
}
\value{
An spplot.
}
\description{
Plots an interpolated surface with the radiocarbon ages of archaeological
sites.
}
