% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpand.R
\name{plot.xpanDates}
\alias{plot.xpanDates}
\title{S3 method for plotting classified archaeological sites and dates of ExPaND
project. It is possible to plot only the sites according to their cultural
affiliation or an interpolated surface with the radiocarbon ages.}
\usage{
plot.xpanDates(sites, culture = "all", isochrones = FALSE)
}
\arguments{
\item{sites}{An xpanDates object.}

\item{culture}{A string. The archaeological culture to be plotted. One of
"all", "BB" (Bacabal and related), "CC" (Cumancaya and related), "GM"
(Goya-Malabrigo), "HZ" (Zone-Hachured), "IP" (Incised-Punctate and related),
"NI" (Unclassified), "PC" (Pedra do Caboclo/Aratu), "PL" (Amazon Polychrome),
"SB" (Saladoid-Barrancoid and related), "TP" (Tupiguarani), "TT"
(Tutishcainyo and related) or "UN" (Una/Taquara/Itararé). Default is "all".}

\item{isochrones}{A boolean indicating whether to plot an interpolated
surface of radiocarbon ages. Default is FALSE.}
}
\value{
An spplot.
}
\description{
S3 method for plotting classified archaeological sites and dates of ExPaND
project. It is possible to plot only the sites according to their cultural
affiliation or an interpolated surface with the radiocarbon ages.
}
