% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpand.R
\docType{data}
\name{sam}
\alias{sam}
\title{South American country borders from NaturalEarth.}
\format{A SpatialPolygonsDataFrame.}
\usage{
sam
}
\description{
South American country borders from NaturalEarth.
}
\keyword{datasets}
