% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpand.R
\docType{data}
\name{xpand}
\alias{xpand}
\title{Radiocarbon dates and coordinates of 1035 archaeological sites in lowland
South America associated with the spread of ceramics and tropical forest
farming.}
\format{A SpatialPointsDataFrame with 2794 features and 11 variables.
\itemize{
  \item Site. Site name.
  \item C14Age. Date in C14 years BP.
  \item C14SD. Standard error of the radiocarbon date.
  \item LabCode. Laboratory code of the C14 date.
  \item Material. Material dated (Charcoal, shell etc.).
  \item Culture. Cultural affiliation (Saladoid, Incised-Punctate,
                 Tupiguarani etc).
  \item Description. Summary description of the archaeological site.
  \item Comments. Comments mainly about problems with the C14 date.
  \item Reference. Author-date reference for the C14 date.
  \item FullReference. Full bibliographic reference.
  \item Exclude. TRUE or FALSE based on general consensus or problems with
                 the C14 date.
}}
\usage{
xpand
}
\description{
Radiocarbon dates and coordinates of 1035 archaeological sites in lowland
South America associated with the spread of ceramics and tropical forest
farming.
}
\keyword{datasets}
